Implementacion de consultas, segunda tarea Agustin Canteros

+++ src/main/java/com/api/productengine/repository/ProductRepository.java
@@ -18,7 +18,7 @@ public interface ProductRepository extends JpaRepository<Product, Long> {
 
     // 1. JPQL with Named Parameters
     @Query("SELECT p FROM Product p WHERE p.name LIKE %:keyword% AND p.price <= :maxPrice")
-    List<Product> searchProducts(@Param("keyword") String keyword, @Param("maxPrice") Double maxPrice);
+    List<Product> searchProducts(@Param("keyword") String keyword, @Param("maxPrice") BigDecimal maxPrice);
 
     // 2. Native Query (Standard SQL)
     @Query(value = "SELECT SUM(price * stock) FROM products", nativeQuery = true)
